kubelet_labels: fix adding default worker node label
The following worker node label `"node.kubernetes.io/node" = ""` show
always be added and appended along with provided labels in lokocfg, as
`TestSelfHostedKubeletLabels` checks the presence of this label and
determines it as a worker node.

In the current incarnation the label is only added if no labels are
provided in lokocfg file. This has been changed to reflect the above
behaviour.

Signed-off-by: Imran Pochi <imran@kinvolk.io>