baremetal: update the kernel arguments when skipping a PXE boot
When the PXE boot is skipped, the new Ignition config is already
fetched but it does not contain the kernel arguments because these
have to be set outside of Ignition or otherwise they would only
apply to the following boots. The kernel arguments are kept in the
grub config and this needs to be updated out of band (kernel console
and Ignition URL are also part of the kernel arguments).