test/components/kubernetes: use more robust label selector
This commit updates `TestSelfHostedKubeletLabels` w.r.t to the new
changes of worker node labels. Previously the label
`node.kubernetes.io/node=` was added as part of the worker node kubelet
label by default. Since the move to using generic controller and worker
modules, this is not true. The above label gets appended to the kubelet
labels only if no other labels are passed by the user.

This change required the test to be updated on how the test lists and
filters the worker nodes. Instead of using the above label, we now list
worker nodes by filtering for the absence of the label
`node.kubernetes.io/controller` which should only be present on
controller nodes.

Signed-off-by: Imran Pochi <imran@kinvolk.io>