assets/terraform-modules/node: remove host_dns_ip
This commit removes the `host_dns_ip` variable from the node template
and removes from controller and worker modules.

Instead of being a part of the node template, this can now be passed as
a CLC snippet.

It is removed becuase, if other platforms start using this module,
presence of setting the host_dns_ip creates `8.8.8.8` as the first
nameserver entry in `/etc/resolv.conf`. This may be undesirable for
platforms such as baremetal which depend on dnsmasq for dns.

This change is needed in Tinkerbell due to limitations of Tinkerbell,
cause by https://github.com/tinkerbell/tink/issues/439.

Signed-off-by: Imran Pochi <imran@kinvolk.io>