baremetal: splitting up matchbox into a separate module.
This commit restructures the baremetal terraform module to split the
flatcar provisioning as a separate module `matchbox-flatcar`.

The new module takes care of setting up matchbox profiles and groups.

This makes it much easier to use this new module for other projects and
possibly the matchbox-flatcar module could maybe live in its own
repository in the future.

Signed-off-by: Imran Pochi <imran@kinvolk.io>