controller/worker: `set_standard_hostname` variable to set hostname
This commit introduces a boolean variable `set_standard_variable` to
controller and worker modules. If set to true this would add the
hostname to the `/etc/hostname` in the node template.

Also adds `cluster_name` variable to the worker module, to construct the
hostname.

With the help of `set_standard_hostname` variable setting the hostname
becomes part of the node template as a condition rather than adding as a
clc snippet.

If set to true this appends the node template with the clc snippet to
create `/etc/hostname` file with the hostname provided to the template.

For controllers, hostname  is set to `<cluster-name>-controller-<index>`
For workers, hostname  is set to `<cluster-name>-worker-<index>`.

Signed-off-by: Imran Pochi <imran@kinvolk.io>