apiserver: Use private key for `service-account-key-file`
For the apiserver flag `--service-account-key-file` we use to provide
public service-account key. Now this commit provides the private
service-account key, the same as the one provided to
`--service-account-signing-key-file`.

Since there is no need of the public key this commit also removes the
whole pipeline that embedded the public key into the kube-apiserver
secret.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>