AKS CI: Increase the timeout for metrics test
This commit increases the timeout of tests with build tag `poste2e` from
the default 10 minutes to 17 minutes and the PollImmediate timeout from
9 minutes to 16 minutes.

On the AKS platform there are not a lot of components installed and the
general e2e tests complete really fast even before prometheus-operator
has discovered the ServiceMonitors in other namespaces and allowed
Prometheus any time to scrape these components. So just the timeout of 9
minutes for polling function and 10 minutes for the entire test suite is
not enough.

In the trial and errors it was discovered that AKS CI needs at least 14
minutes to do the discovery.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>