matchbox-flatcar: disable all LVM mappings
Since the disks may or not be not empty, any detected LVM partitions may
automatically be activated on bootup. This can disrupt the installation
and we need to disable them first. This was done for all "vg" prefixes
but we can do it for any volume names. Any non-LVM entries are ignored
with a small error log output.

See https://github.com/kinvolk/Flatcar/issues/332