matchbox-flatcar: wipe additional disks
Exposes a variable to the user for wiping any additional disks attached
to the machine. Default is false.

When set to true, any additional disks attached to the hardware are
wiped clean before proceeding with the Flatcar Container Linux
installation.

The reason this is being done is to provide clean and unformatted disk
devices to the Lokomotive components [openebs and rook] for storage
orchestration.

Both rook and openebs fail to use the disks that have data or filesystem
already present on the disk. This leads to among other issues, failed
installation where PVCs get stuck in pending state.

Signed-off-by: Imran Pochi <imran@kinvolk.io>