baremetal: introduce CLC snippets for the installer
The installer OS needs to be able to get its network configuration as
CLC snippet if DHCP is not used.

Add CLC snippets for the installer and migrate the matchbox-specific
CLC template to the normal CLC template which is needed for CLC
snippets. The request template variable is replaced by the actually
used "mac" parameter.