baremetal: use ignition.config.url kernel parameter
The Ignition configuration currently gets written out only once to the
OEM partition and any updates to it in the matchbox store are not
propagated to the node's Ignition environment when Ignition is forced
to run again.

Use the Ignition kernel parameter to point the node's Ignition
environment directly to matchbox where the userdata lies.