aws ci: Use c5a.2xlarge for controllers
AWS CI keeps failing on cert-rotation test because apiserver takes a
really long time to come back up.

It is possible that the failure rate on AWS is higher because the
machine type on AWS is tiny which is unable to sustain the load
generated due to control plane component restarts.

This commit upgrades the control plane machine to fairly larger one.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>