Only restart the containers from systemd
Restarting etcd or the kubelet container from the Docker daemon
introduces a race because the "docker logs" command will terminate and
then the systemd unit will also restart the container in parallel to
the Docker daemon. This race causes problems like terminating the
container again or failing to start it, but it also introduces
additional issues because the ExecStartPre and ExecStopPost commands
run while the container is running, which means the state can get
corrupted. Even if the systemd unit would not restart in parallel, it
is not wanted to restart the container from the Docker daemon because
the systemd unit should first run the ExecStartPre/Post commands, e.g.,
for the "etcd-rejoin" script.

Only restart the containers from the systemd unit, not from the Docker
daemon.