etcd: stop the service before modifying the etcd folder
The etcd folder was modified while the etcd service was running which
could lead to an earlier restart of the service because the service
conditions for existing files are finally met and it may conflict
with any commands in the ExecStopPost directive.

First stop the etcd service unit before touching the files and only
start the unit after the modification is done.