aws|packet: align etcd unit file with the shared controller module
The shared etcd CLC systemd unit has additional conditions for startup
which were missing in the aws and packet etcd systemd units. It seems
that these conditions help to make the bootstap more robust.

Port the startup conditions over to aws and packet to align the
platforms (which also makes debugging more clear).