terraform: remove kubeconfig ConditionPathExists on kubelet.service
It's added to the [Service] section which means it has no effect.
Adding it to the right place (the [Unit] section) means
/etc/kubernetes/kubeconfig must exist the first time the Kubelet starts,
otherwise the unit is skipped and it never tries to start again.

This removes that directive: it's fine to have the kubelet try to start
again and again until the file exists. Note given the current
configuration (RestartSec=5 and the default StartLimitIntervalSec=10s
and StartLimitBurst=5) backoff will never be triggered so it will
restart indefinitely.