feat: define MaxHistory for Helm releases
Defining the `MaxHistory` option during the rollback Helm action so that
the releases history is cleaned up accordingly.
Without this option the history would never been cleared which is
accumulating old configuration which can be harmful when a incorrect
rollback is applied. By cleaning up the history, we limit the potential
impact rolling back to a older version.