baremetal: add ignore_worker_changes option
The baremetal platform defaults to replacing worker nodes on changes
through a simple reboot with rerunning Ignition.
This does not work for pre-0.9 deployed workers because they don't
use the ignition kernel flag to fetch the configuration. These nodes
either need a PXE boot for reprovisioning or, if the user wants to
keep the state, an additional variable to prevent them being replaced.
This additional variable is also useful if users have a workload that
doesn't play well with replacing the worker.

Add a new variable to ignore worker config changes, similar to what
is already done on Equinix Metal, but keep replacing the workers on
by default.